+++ commanderbot_lib/database/yaml_versioned_file_database.py
@@ -1,10 +1,10 @@
 from typing import IO
 
-from commanderbot_lib.database.abc.file_database import FileDatabase
+from commanderbot_lib.database.abc.versioned_file_database import VersionedFileDatabase
 from commanderbot_lib.database.mixins.yaml_file_database_mixin import YamlFileDatabaseMixin
 
 
-class YamlFileDatabase(FileDatabase, YamlFileDatabaseMixin):
+class YamlVersionedFileDatabase(VersionedFileDatabase, YamlFileDatabaseMixin):
     # @implements FileDatabase
     async def load(self, file: IO) -> dict:
         return await self.load_yaml(file)