+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [Unreleased]
+
+### Added
+
+- Implemented file database backups for `FileDatabase` implementations
+  - The backup file will use the original name with a microsecond-accurate timestamp inserted before the file extension
+  - `VersionedFileDatabase` now creates a backup copy of the database file before applying any writing-back any changes
+
 ## [0.5.0] - 2021-01-03
 
 ### Changed