+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - The backup file will use the original file name with a microsecond-accurate timestamp inserted before the file extension
   - `VersionedFileDatabase` now creates a backup copy of the database file before writing-back any changes applied by migrations
 
+### Changed
+
+- Updated to discord.py 1.6
+
 ## [0.5.0] - 2021-01-03
 
 ### Changed